requirements.txt
aiohttp==3.6.2
amqp==2.5.2
async-timeout==3.0.1
attrs==19.1.0
billiard==3.6.0.0
bottle==0.12.17
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.3.0
celery-once==3.0.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
configobj==5.0.6
cryptography==2.8
cssselect2==0.2.1
dbus-python==1.2.8
defusedxml==0.6.0
discord.py==1.5.1
Django==2.2.2
django-admin-sortable2==0.7.2
django-admin-views==0.8.0
django-bootstrap4==0.0.8
django-computedfields==0.0.10
django-dynamic-preferences==1.7
django-environ==0.4.5
django-extensions==2.1.7
django-redis==4.10.0
future==0.18.2
html5lib==1.0.1
idna==2.8
kombu==4.6.1
multidict==4.5.2
persisting-theory==0.2.1
Pillow==6.1.0
psycopg2-binary==2.8.4
pycparser==2.19
pycryptodome==3.9.8
pycurl==7.43.0.5
pyotp==2.3.0
Pyphen==0.9.5
python-systemd==0.0.9
python-telegram-bot==12.2.0
pytz==2019.1
qrcode==6.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
tornado==6.0.3
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
websockets==6.0
yarl==1.3.0
